test: stabilize subscription import coverage
